requirements.txt
streamlit
rasa
nltk
google-generativeai
